Extract HttpBase class
