+++ lib/firefly.ts
@@ -1,9 +1,6 @@
 import { Stream, Readable } from 'stream';
-import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import * as FormData from 'form-data';
 import {
-  FireFlyOptions,
-  FireFlyOptionsInput,
   FireFlyStatusResponse,
   FireFlyPrivateSendOptions,
   FireFlyCreateOptions,
@@ -51,106 +48,10 @@ import {
   FireFlyContractAPIFilter,
 } from './interfaces';
 import { FireFlyWebSocket, FireFlyWebSocketCallback } from './websocket';
+import HttpBase, { mapConfig } from './http';
 
-function isSuccess(status: number) {
-  return status >= 200 && status < 300;
-}
-
-function mapConfig(
-  options: FireFlyGetOptions | FireFlyCreateOptions | undefined,
-  params?: any,
-): AxiosRequestConfig {
-  return {
-    ...options?.requestConfig,
-    params: {
-      ...params,
-      confirm: options?.confirm,
-    },
-  };
-}
-
-export class FireFlyError extends Error {}
-
-export default class FireFly {
-  private options: FireFlyOptions;
-  private rootHttp: AxiosInstance;
-  private http: AxiosInstance;
+export default class FireFly extends HttpBase {
   private queue = Promise.resolve();
-  private errorHandler?: (err: FireFlyError) => void;
-
-  constructor(options: FireFlyOptionsInput) {
-    this.options = this.setDefaults(options);
-    this.rootHttp = axios.create({
-      ...options.requestConfig,
-      baseURL: `${options.host}/api/v1`,
-    });
-    this.http = axios.create({
-      ...options.requestConfig,
-      baseURL: `${options.host}/api/v1/namespaces/${this.options.namespace}`,
-    });
-  }
-
-  private setDefaults(options: FireFlyOptionsInput): FireFlyOptions {
-    return {
-      ...options,
-      namespace: options.namespace ?? 'default',
-      websocket: {
-        ...options.websocket,
-        host: options.websocket?.host ?? options.host.replace('http', 'ws'),
-        reconnectDelay: options.websocket?.reconnectDelay ?? 5000,
-        heartbeatInterval: options.websocket?.heartbeatInterval ?? 30000,
-      },
-    };
-  }
-
-  private async wrapError<T>(response: Promise<AxiosResponse<T>>) {
-    return response.catch((err) => {
-      if (axios.isAxiosError(err)) {
-        const errorMessage = err.response?.data?.error;
-        const ffError = new FireFlyError(errorMessage ?? err.message);
-        if (this.errorHandler !== undefined) {
-          this.errorHandler(ffError);
-        }
-        throw ffError;
-      }
-      throw err;
-    });
-  }
-
-  private async getMany<T>(url: string, params?: any, options?: FireFlyGetOptions, root = false) {
-    const http = root ? this.rootHttp : this.http;
-    const response = await this.wrapError(http.get<T>(url, mapConfig(options, params)));
-    return response.data;
-  }
-
-  private async getOne<T>(url: string, options?: FireFlyGetOptions, params?: any, root = false) {
-    const http = root ? this.rootHttp : this.http;
-    const response = await this.wrapError(
-      http.get<T>(url, {
-        ...mapConfig(options, params),
-        validateStatus: (status) => status === 404 || isSuccess(status),
-      }),
-    );
-    return response.status === 404 ? undefined : response.data;
-  }
-
-  private async createOne<T>(url: string, data: any, options?: FireFlyCreateOptions) {
-    const response = await this.wrapError(this.http.post<T>(url, data, mapConfig(options)));
-    return response.data;
-  }
-
-  private async replaceOne<T>(url: string, data: any) {
-    const response = await this.wrapError(this.http.put<T>(url, data));
-    return response.data;
-  }
-
-  private async deleteOne<T>(url: string) {
-    await this.wrapError(this.http.delete<T>(url));
-  }
-
-  onError(handler: (err: FireFlyError) => void) {
-    this.errorHandler = handler;
-  }
 
   async getStatus(options?: FireFlyGetOptions): Promise<FireFlyStatusResponse> {
     const response = await this.rootHttp.get<FireFlyStatusResponse>('/status', mapConfig(options));
