+++ lib/index.ts
@@ -1,5 +1,3 @@
 import FireFly from './firefly';
-
 export default FireFly;
-export * from './firefly';
 export * from './interfaces';
